updating fake TOAs.  Moving to a new module
