+++ src/pint/scripts/zima.py
@@ -8,22 +8,12 @@ from astropy.time import TimeDelta
 import pint.fitter
 import pint.models
 import pint.toa as toa
-from pint.observatory import get_observatory
+import pint.simulation
+import pint.residuals
 
 __all__ = ["main"]
 
 # log.setLevel("INFO")
-
-
-def get_freq_array(base_freq_values, ntoas):
-    """Right now it is a very simple frequency array simulation.
-       It just simulates an alternating frequency arrays
-    """
-    freq = np.zeros(ntoas)
-    num_freqs = len(base_freq_values)
-    for ii, fv in enumerate(base_freq_values):
-        freq[ii::num_freqs] = fv
-    return freq
 
 
 def main(argv=None):
@@ -65,6 +55,12 @@ def main(argv=None):
         default=1.0,
     )
     parser.add_argument(
+        "--addnoise",
+        action="store_true",
+        default=False,
+        help="Actually add in random noise, or just populate the column",
+    )
+    parser.add_argument(
         "--fuzzdays",
         help="Standard deviation of 'fuzz' distribution (jd) (default: 0.0)",
         type=float,
@@ -72,13 +68,6 @@ def main(argv=None):
     )
     parser.add_argument(
         "--plot", help="Plot residuals", action="store_true", default=False
-    )
-    parser.add_argument("--ephem", help="Ephemeris to use", default="DE421")
-    parser.add_argument(
-        "--planets",
-        help="Use planetary Shapiro delay",
-        action="store_true",
-        default=False,
     )
     parser.add_argument(
         "--format", help="The format of out put .tim file.", default="TEMPO2"
@@ -93,59 +82,28 @@ def main(argv=None):
 
     if args.inputtim is None:
         log.info("Generating uniformly spaced TOAs")
-        duration = args.duration * u.day
-        # start = Time(args.startMJD,scale='utc',format='pulsar_mjd',precision=9)
-        start = np.longdouble(args.startMJD) * u.day
-        freq = np.atleast_1d(args.freq) * u.MHz
-        site = get_observatory(args.obs)
-        scale = site.timescale
-
-        times = np.linspace(0, duration.to(u.day).value, args.ntoa) * u.day + start
-
-        # 'Fuzz' out times
-        if args.fuzzdays > 0.0:
-            fuzz = np.random.normal(scale=args.fuzzdays, size=len(times)) * u.day
-            times += fuzz
-
-        # Add mulitple frequency
-        freq_array = get_freq_array(freq, len(times))
-        tl = [
-            toa.TOA(t.value, error=error, obs=args.obs, freq=f, scale=scale)
-            for t, f in zip(times, freq_array)
-        ]
-        ts = toa.TOAs(toalist=tl)
+        ts = pint.simulation.make_fake_toas_uniform(
+            startMJD=args.startMJD,
+            endMJD=args.startMJD + args.duration,
+            ntoas=args.ntoa,
+            model=m,
+            obs=args.obs,
+            error=error,
+            freq=np.atleast_1d(args.freq) * u.MHz,
+            fuzz=args.fuzzdays * u.d,
+            add_noise=args.addnoise,
+        )
     else:
         log.info("Reading initial TOAs from {0}".format(args.inputtim))
-        ts = toa.TOAs(toafile=args.inputtim)
-        ts.table["error"][:] = error
-
-    # WARNING! I'm not sure how clock corrections should be handled here!
-    # Do we apply them, or not?
-    if not any(["clkcorr" in f for f in ts.table["flags"]]):
-        log.info("Applying clock corrections.")
-        ts.apply_clock_corrections()
-    if "tdb" not in ts.table.colnames:
-        log.info("Getting IERS params and computing TDBs.")
-        ts.compute_TDBs(ephem=args.ephem)
-    if "ssb_obs_pos" not in ts.table.colnames:
-        log.info("Computing observatory positions and velocities.")
-        ts.compute_posvels(args.ephem, args.planets)
-
-    log.info("Creating TOAs")
-    F_local = m.d_phase_d_toa(ts)
-    rs = m.phase(ts).frac / F_local
-
-    # Adjust the TOA times to put them where their residuals will be 0.0
-    ts.adjust_TOAs(TimeDelta(-1.0 * rs))
-    rspost = m.phase(ts).frac / F_local
-
-    log.info("Second iteration")
-    # Do a second iteration
-    ts.adjust_TOAs(TimeDelta(-1.0 * rspost))
-
-    err = np.random.randn(len(ts.table)) * error
-    # Add the actual error fuzzing
-    ts.adjust_TOAs(TimeDelta(err))
+        ts = pint.simulation.make_fake_toas_fromtim(
+            args.inputtim,
+            model=m,
+            obs=args.obs,
+            error=error,
+            freq=np.atleast_1d(args.freq) * u.MHz,
+            fuzz=args.fuzzdays * u.d,
+            add_noise=args.addnoise,
+        )
 
     # Write TOAs to a file
     ts.write_TOA_file(args.timfile, name="fake", format=out_format)
@@ -157,16 +115,13 @@ def main(argv=None):
 
         quantity_support()
 
-        rspost2 = m.phase(ts).frac / F_local
+        r = pint.residuals.Residuals(ts, m)
         plt.errorbar(
-            ts.get_mjds(), rspost2.to(u.us), yerr=ts.get_errors().to(u.us), fmt="."
+            ts.get_mjds(),
+            r.calc_time_resids().to(u.us),
+            yerr=ts.get_errors().to(u.us),
+            fmt=".",
         )
-        newts = pint.toa.get_TOAs(args.timfile, ephem=args.ephem, planets=args.planets)
-        rsnew = m.phase(newts).frac / F_local
-        plt.errorbar(
-            newts.get_mjds(), rsnew.to(u.us), yerr=newts.get_errors().to(u.us), fmt="."
-        )
-        # plt.plot(ts.get_mjds(),rspost.to(u.us),'x')
         plt.xlabel("MJD")
         plt.ylabel("Residual (us)")
         plt.grid(True)
